% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-G.R
\name{pdf_theorical}
\alias{pdf_theorical}
\title{MOTP-Weibull density function (MOTPW)}
\usage{
pdf_theorical(x, alpha, theta, beta, lambda)
}
\arguments{
\item{x}{support of the probability density function.}

\item{alpha}{Parameter.}

\item{theta}{Parameter.}

\item{beta}{Parameter.}

\item{lambda}{Parameter.}
}
\description{
MOTP-Weibull density function (MOTPW)
}
\examples{

# Checking the implementation of the theorical density function:
integrate(
  pdf_theorical,
  lower = 0,
  upper = 20,
  alpha = 1,
  theta = 0.5,
  beta = 1,
  lambda = 1
)
}
