% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_series_distributions.R
\name{pf_geometric}
\alias{pf_geometric}
\title{Geometric probability function}
\usage{
pf_geometric(n, theta)
}
\arguments{
\item{n}{Function support, with \eqn{n = 1, 2, \cdots}.}

\item{theta}{A probability value, i.e, \eqn{0 < \theta < 1}.}
}
\value{
Calculated probability (vector)
}
\description{
Geometric probability function
}
\examples{
sapply(
   X = 1L:1000L,
   FUN = function(i)
   pf_geometric(n = i, theta = 0.5)
) \%>\% sum
}
