% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_series_distributions.R
\name{pf_ztp}
\alias{pf_ztp}
\title{Zero-truncated Poisson probability function}
\usage{
pf_ztp(n, theta)
}
\arguments{
\item{n}{Function support, with \eqn{n = 1, 2, \cdots}.}

\item{theta}{Theta parameter, such that \eqn{\theta > 0}.}
}
\value{
Calculated probability (vector)
}
\description{
Zero-truncated Poisson probability function
}
\examples{
sapply(
   X = 1L:1000L,
   FUN = function(i)
   pf_ztp(n = i, theta = 1.2)
) \%>\% sum
}
