% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-G.R
\name{upsilon}
\alias{upsilon}
\title{Mathematical term used in the sum of equation 19 of the paper}
\usage{
upsilon(n, k, alpha, theta, p_n)
}
\arguments{
\item{n}{Value that indexes a sum of equation 19.}

\item{k}{Value that indexes a sum of equation 19.}

\item{alpha}{Additional parameters introduced by the new model proposed in the paper.}

\item{theta}{Additional parameters introduced by the new model proposed in the paper.}
}
\value{
Return of expression  \eqn{\upsilon_{n,k}(\alpha, \beta)} ot the paper.
}
\description{
Mathematical term used in the sum of equation 19 of the paper
}
\details{
\eqn{n} and \eqn{k} are indexed values in the sums of equation
19 of the paper.
}
