% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-G.R
\name{eq_19}
\alias{eq_19}
\title{Equation 19 of the paper}
\usage{
eq_19(x, G, p_n, N = 100L, K = 100L, alpha, theta, ...)
}
\description{
Equation 19 of the paper
}
\examples{
cdf_w <- function(x, a, b) {
  pweibull(q = x, shape = a, scale = b)
}

eq_19(
  x = 1,
  G = cdf_w,
  p_n = pf_ztp,
  N = 100L,
  K = 100L,
  alpha = 1.2,
  theta = 1.3,
  a = 1,
  b = 1
)

integrate(f = eq_19, lower = 0, upper = 10,
          G = cdf_w, p_n = pf_ztp, N = 100L, K = 100L,
          alpha = 1.2, theta = 1.3, a = 1, b = 1)
}
