% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-G.R
\name{eq_19}
\alias{eq_19}
\title{Equation 19 of the paper}
\usage{
eq_19(x, G, p_n, N = 50L, K = 50L, alpha, theta, ...)
}
\description{
Equation 19 of the paper
}
\examples{
# Theoretical density (eq. 9 of the paper):
pdf_motpw <- function(x, alpha, theta, beta, lambda) {
   u <- (lambda * x)^beta

   part_1 <- (alpha * theta * beta * lambda^beta * x^(beta - 1) * exp(-u)) /
   ((exp(theta) - 1) * (1 - (1 - alpha) * exp(-u))^2)

   part_2 <- exp(theta * (1 - exp(-u)) / (1 - (1 - alpha) * exp(-u)))

   part_1 * part_2
}

# Arriving the value of the integral:
integrate(
  pdf_motpw,
  lower = 0,
  upper = 20,
  alpha = 1,
  theta = 0.5,
  beta = 1,
  lambda = 1
)

##########################

cdf_w <- function(x, beta, lambda) {
  pweibull(q = x, shape = beta, scale = lambda)
}

f <- function(x) {
   eq_19(
      x = x,
      G = cdf_w,
      p_n = pf_ztp,
      N = 200L,
      K = 100L,
      alpha = 1.2,
      theta = 0.3,
      beta = 1,
      lambda = 1
    )
}

f(x = 0.5)
}
