% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-G.R
\name{eq_19}
\alias{eq_19}
\title{Equation 19 of the paper}
\usage{
eq_19(x, G, p_n, N = 100L, K = 100L, alpha, theta, ...)
}
\description{
Equation 19 of the paper
}
\examples{
cdf_w <- function(x, a, b) {
   pweibull(q = x, shape = a, scale = b)
}
eq_19(alpha, theta, p_n, G)
}
