% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-plot.R
\name{eval_plot_moptw}
\alias{eval_plot_moptw}
\title{Plot Eval MOPTW}
\usage{
eval_plot_moptw(
  n_points = 5L,
  N = 50L,
  K = 50L,
  xmin,
  xmax,
  alpha,
  theta,
  beta,
  lambda
)
}
\description{
Plot Eval MOPTW
}
\examples{
eval_plot_moptw(
   n_points = 5L,
   N = 100L,
   K = 100L,
   xmin = 0.01,
   xmax = 2,
   alpha = 0.5,
   theta = 0.5,
   beta = 0.5,
   lambda = 0.5
)
}
